If testing, add the itag to the end of the list before sorting.
Python 3 changed the way sorting works for functionally identical
objects. We add the itag at the end to make them unique so that the
sorting order is always the same, which means we can easily test that
the function works as expected. This is not required in normal use,
because we don't care which order identical itags are returned in
relative to each other.